Added settings for group level, and project level. Added option for purgingIssues everytime, or not. Fix gitlab PAT reloading plugin. Fix gitlab icon.

+++ src/gitlab-loader.ts
@@ -14,13 +14,27 @@ export default class GitlabLoader {
 		this.settings = settings;
 	}
 
+	getUrl() {
+		switch (this.settings.gitlabIssuesLevel) {
+			case "project":
+				return `${this.settings.gitlabApiUrl()}/projects/${this.settings.gitlabAppId}/issues?${this.settings.filter}`;
+			case "group":
+				return `${this.settings.gitlabApiUrl()}/groups/${this.settings.gitlabAppId}/issues?${this.settings.filter}`;
+			case "personal":
+			default:
+				return `${this.settings.gitlabApiUrl()}/issues?${this.settings.filter}`;
+		}
+	}
+
 	loadIssues() {
-		// URL Encode the user filter settings
-		const apiURL = `${this.settings.gitlabApiUrl()}/issues?${this.settings.filter}`;
-		GitlabApi.load<Array<Issue>>(encodeURI(apiURL), this.settings.gitlabToken)
+		GitlabApi.load<Array<Issue>>(encodeURI(this.getUrl()), this.settings.gitlabToken)
 			.then((issues: Array<Issue>) => {
 				const gitlabIssues = issues.map((rawIssue: Issue) => new GitlabIssue(rawIssue));
-				this.fs.purgeExistingIssues();
+
+				if(this.settings.purgeIssues) {
+					this.fs.purgeExistingIssues();
+				}
+
 				this.fs.processIssues(gitlabIssues);
 			})
 			.catch(error => {
